+++ frontend-tests/tests/2_check_rules.js
@@ -3,21 +3,42 @@
 import { Selector } from 'testcafe'
 import { login } from '../utils/logins.js'
 import baseurl from '../utils/url.js'
-import { leadZero, randNum } from '../utils/utils.js'
+import { leadZero, randNum, createDate } from '../utils/utils.js'
+import { createCase, createAppropriation, createActivity, approveActivities } from '../utils/crud.js'
 
 const today = new Date(),
     anotherday = new Date(new Date().setDate(today.getDate() + 2)),
     testdata = {
-        case: {
+        case1: {
             name: `${ randNum() }.${ randNum() }.${ randNum() }-regel-test`
         },
-        appr: {
+        case2: {
+            name: `${ randNum() }.${ randNum() }.${ randNum() }-regel-test`
+        },
+        appr1: {
             name: `${ randNum() }.${ randNum() }.${ randNum() }-regel-test-bevilling`
         },
-        act: {
+        appr2: {
+            name: `${ randNum() }.${ randNum() }.${ randNum() }-regel-test-bevilling`
+        },
+        act1: {
             start_today: `${ today.getFullYear() }-${ leadZero(today.getMonth() + 1) }-${ leadZero(today.getDate()) }`,
             end_date: `${ anotherday.getFullYear() }-${ leadZero(anotherday.getMonth() + 1) }-${ leadZero(anotherday.getDate()) }`
+        },
+        act2: {
+            type: 1,
+            start: createDate(3),
+            amount: '12',
+            payee_id: '89837837728',
+            payee_name: 'Fiktiv Virksomhed I/S'
+        },
+        act3: {
+            expected_type: 'adjustment',
+            type: 1,
+            start: createDate(5),
+            amount: '15'
         }
+
     }
 
 fixture `Check payments start date`
@@ -28,33 +49,37 @@ fixture `Check payments start date`
 
 test('Start date rule check for activities with "CASH" payment', async t => {
 
+    await createCase(t, testdata.case1)
+    await createAppropriation(t, testdata.appr1)
+
     await t
-        .click(Selector('button').withText('+ Tilknyt hovedsag'))
-        .typeText('#field-sbsys-id', testdata.case.name)
-        .typeText('#field-cpr', '000000-0000')
-        .click(Selector('label').withAttribute('for', 'inputRadio1'))
-        .click('#selectField4')
-        .click(Selector('#selectField4 option').withText('Baltorp'))
-        .click('#field-indsatstrappe')
-        .click(Selector('#field-indsatstrappe option').withText('Trin 3: Hjemmebaserede indsatser'))
-        .click('#field-skaleringstrappe')
-        .click(Selector('#field-skaleringstrappe option').withText('5'))
-        .click(Selector('input').withAttribute('type', 'submit'))
-        .click(Selector('.appropriation-create-btn'))
-    
-        .typeText('#field-sbsysid', testdata.appr.name)
-        .click('#field-lawref')
-        .click(Selector('#field-lawref option').nth(1))
-        .click(Selector('input').withAttribute('type', 'submit'))
         .click(Selector('.activities-create-btn'))
         .click('#fieldSelectAct')
         .click(Selector('#fieldSelectAct option').nth(1))
-        .typeText('#field-startdate', testdata.act.start_today)
+        .typeText('#field-startdate', testdata.act1.start_today)
         .click('#field-payee')
         .click(Selector('#field-payee option').nth(2))
         .click('#field-pay-method')
         .click(Selector('#field-pay-method option').nth(0))
         .expect(Selector('.warning').exists).ok()
-        .typeText('#field-startdate', testdata.act.end_date)
+        .typeText('#field-startdate', testdata.act1.end_date)
         .expect(Selector('.warning').exists).notOk()
+})
+
+test('Check that an activity may only have one modifier', async t => {
+
+    await createCase(t, testdata.case2)
+    await createAppropriation(t, testdata.appr2)
+    await createActivity(t, testdata.act2)
+    await approveActivities(t)
+    
+    await t.click(Selector('.act-list-item a'))
+
+    await createActivity(t, testdata.act3)
+
+    await t
+        .click(Selector('.meta-row'))
+        .click(Selector('.sub-row a'))
+        .expect(Selector('button').withText('Lav forventet justering').exists).notOk()
+
 })